# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/index": {
            "get": {
                "tags": [
                    "首頁"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/createUser": {
            "get": {
                "tags": [
                    "用戶"
                ],
                "summary": "新增用戶",
                "parameters": [
                    {
                        "type": "string",
                        "description": "使用者名稱",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "使用者密碼",
                        "name": "password",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "使用者確認密碼",
                        "name": "repassword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/deleteUser": {
            "get": {
                "tags": [
                    "用戶"
                ],
                "summary": "刪除用戶",
                "parameters": [
                    {
                        "type": "string",
                        "description": "使用者ID",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/findUserByNameAndPwd": {
            "post": {
                "tags": [
                    "用戶"
                ],
                "summary": "所有用戶",
                "parameters": [
                    {
                        "type": "string",
                        "description": "使用者名稱",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "使用者密碼",
                        "name": "password",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/getUserList": {
            "get": {
                "tags": [
                    "用戶"
                ],
                "summary": "所有用戶",
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/updateUser": {
            "post": {
                "tags": [
                    "用戶"
                ],
                "summary": "編輯用戶",
                "parameters": [
                    {
                        "type": "string",
                        "description": "使用者ID",
                        "name": "id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "使用者名稱",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "使用者密碼",
                        "name": "password",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "電話號碼",
                        "name": "phone",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "電子信箱",
                        "name": "email",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}